+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-08-17 16:12:22
-// Commit: fix(i18n): prevent double tap duplicate action in i18n
+// Updated at: 2026-08-18 11:48:49
+// Commit: test(i18n): implement widget integration tests for i18n
 
 class i18nComponentState {
-  final String commitRef = '3670';
-  final String lastMessage = 'fix(i18n): prevent double tap duplicate action in i18n';
-  final DateTime timestamp = DateTime.parse('2026-08-17T16:12:22+05:00');
+  final String commitRef = '3689';
+  final String lastMessage = 'test(i18n): implement widget integration tests for i18n';
+  final DateTime timestamp = DateTime.parse('2026-08-18T11:48:49+05:00');
 }